Fixing file paths for tests running in Linux containers/envs
Fix loading Credential Manager in Linux containers/envs

Fix loading Env strings on Linux containers/envs

Fix when there is no existing ACME accounts in a container/env

Initial implementation for managing certs on different OSs

Changes for testing error message for bad Cert Store name on various OSs

Improve error logging for Binding Match tests

Changes for getting debug info on non-Windows systems

Changes for making Linux and Windows Docker Images of Core Unit Tests

Only the .NET 8.0 images are currently working as they should, there are some issues with the .NET 4.6.2 images for Windows and Linux

Fix to generating debug info for non-Windows OSs

Create common utility method to get Environment values in Account Tests

Remove unneeded changes to Certify.Core.Tests.Unit.csproj

Improve path strings for multiple OSs using Path.Combine()

Changes to Linux Dockerfiles for Certify.Core.Tests.Unit

Changes to enable using Step CA for CertifyManagerAccount unit tests

Fix to BindingMatchTests for bad file path

Windows docker changes

More Windows Docker changes

Fixes to compose files

Documents for running the Certify Core Unit Tests in Docker

Fix to formatting of Certify Core Unit Test Docker.md

Cleanup linux_compose.yaml

+++ src/Certify.Tests/Certify.Core.Tests.Unit/LoggyTests.cs
@@ -9,12 +9,25 @@ namespace Certify.Core.Tests.Unit
     [TestClass]
     public class LoggyTests
     {
-        private string logFilePath = "C:\\ProgramData\\certify\\Tests\\test.log";
+        private string testsDataPath;
+        private string logFilePath;
 
         [TestInitialize]
         public void TestInitialize()
         {
-            File.Delete(this.logFilePath);
+            testsDataPath = Path.Combine(EnvironmentUtil.GetAppDataFolder(), "Tests");
+            logFilePath = Path.Combine(testsDataPath, "test.log");
+
+            if (!Directory.Exists(testsDataPath))
+            {
+                Directory.CreateDirectory(testsDataPath);
+
+            }
+
+            if (File.Exists(logFilePath))
+            {
+                File.Delete(this.logFilePath);
+            }
         }
 
         [TestCleanup]
@@ -56,10 +69,11 @@ namespace Certify.Core.Tests.Unit
 
             // Trigger an exception error and log it using Loggy.Error()
             var logMessage = "New Loggy Exception Error";
-            var exceptionError = "System.IO.FileNotFoundException: Could not find file 'C:\\ProgramData\\certify\\Tests\\test1.log'.";
+            var badFilePath = Path.Combine(EnvironmentUtil.GetAppDataFolder(), "Tests", "test1.log");
+
+            var exceptionError = $"System.IO.FileNotFoundException: Could not find file '{badFilePath}'.";
             try
             {
-                var badFilePath = "C:\\ProgramData\\certify\\Tests\\test1.log";
                 var nullObject = File.ReadAllBytes(badFilePath);
             }
             catch (Exception e)
